Refactor checkYaml functions
